requirements.txt
tinkoff-investments
load_dotenv
matplotlib
pandas
numpy
ta
asyncio
logging
telegram